internal/parser/parser.go
package parser

import (
	"github.com/matheus-alpe/interpreter/internal/ast"
	"github.com/matheus-alpe/interpreter/internal/lexer"
	"github.com/matheus-alpe/interpreter/internal/token"
)

type Parser struct {
	l *lexer.Lexer

	currentToken token.Token
	peekToken    token.Token
}

func New(l *lexer.Lexer) *Parser {
	p := &Parser{l: l}

	// Read two tokens, so currentToken and peekToken are both set
	p.nextToken()
	p.nextToken()

	return p
}

func (p *Parser) nextToken() {
	p.currentToken = p.peekToken
	p.peekToken = p.l.NextToken()
}

func (p *Parser) ParseProgram() *ast.Program {
	return nil
}
